+++ server/marketplace/src/main/java/com/dspajic/marketplace/mappers/LocationRowMapper.java
@@ -10,7 +10,7 @@ public class LocationRowMapper implements RowMapper<Location> {
     @Override
     public Location mapRow(ResultSet rs, int rowNum) throws SQLException {
         Location e = new Location();
-        e.setId(rs.getInt("location_id"));
+        e.setId(rs.getLong("location_id"));
         e.setAddressLine1(rs.getString("address_line1"));
         e.setAddressLine2(rs.getString("address_line2"));
         e.setPostalCode(rs.getString("postal_code"));
